Move several groups of tasks out of tasks.py
