+++ src/tasks/__init__.py
@@ -1,6 +1,9 @@
 # Import task definitions to register their tasks.
 from . import tasks
 from . import edge_probing
+from . import lm
+from . import mt
+from . import reddit
 
 # REGISTRY needs to be available to modules within this package,
 # but we also import it here to make it available at the package level.
